+++ Classes/Service/Transfer/DatatypeTransferImportService.php
@@ -38,6 +38,7 @@ class DatatypeTransferImportService
         'inline',
         'datatype',
         'dyninput',
+        'repeater',
         'passthrough',
         'tca',
     ];