+++ Classes/Tca/Field/Email.php
@@ -1,0 +1,37 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the package k3n/tonictypes.
+ *
+ * For the full copyright and license information, please read the
+ * LICENSE file that was distributed with this source code.
+ *
+ * Many thanks to Auth: B. Zagar / Maint: J. Pietschmann for sharing this extension – TYPO3 inspiring people to share!
+ * Contact: support@tonictypes.com
+ *
+ */
+
+namespace K3n\Tonictypes\Tca\Field;
+
+/**
+ * Validated email input field (TCA type "email").
+ */
+class Email extends Input
+{
+    public function getTca(): array
+    {
+        $tca = [
+            'exclude' => (int)$this->getField()->isExclude(),
+            'label' => $this->getField()->getFrontendLabel(),
+            'config' => [
+                'type' => 'email',
+                'size' => $this->getField()->getConfig('size') ?: 30,
+                'eval' => 'trim',
+            ],
+        ];
+
+        return $this->mergeConfigurationToTca($tca);
+    }
+}